some repos and services and controllers

+++ core/src/main/java/com/siemens/core/repository/BrokerRepository.java
@@ -2,5 +2,5 @@ package com.siemens.core.repository;
 
 import com.siemens.core.model.Broker;
 
-public interface BrokerRepository extends GenericRepository<Broker, Long> {
+public interface BrokerRepository extends GenericRepository<Broker, Integer> {
 }
