+++ core/src/main/java/com/siemens/core/repository/HoldingRecordRepository.java
@@ -2,5 +2,5 @@ package com.siemens.core.repository;
 
 import com.siemens.core.model.HoldingRecord;
 
-public interface HoldingRecordRepository extends GenericRepository<HoldingRecord, Long> {
+public interface HoldingRecordRepository extends GenericRepository<HoldingRecord, Integer> {
 }
