+++ core/src/main/java/com/siemens/core/repository/UserRepository.java
@@ -2,5 +2,5 @@ package com.siemens.core.repository;
 
 import com.siemens.core.model.User;
 
-public interface UserRepository extends GenericRepository<User, Long> {
+public interface UserRepository extends GenericRepository<User, Integer> {
 }
